requirements.txt
langchain==0.1.8
pinecone-client==2.2.4
python-dotenv==1.0.0
streamlit==1.29.0